Selector v2 line: OnlineSelectorV2, guided BLS v2 with variants v2b-v2g, prequential monitor, calibration experiments and status report
- New selector (vectorised features, ensemble heads + Thompson sampling, Gumbel-top-k orders,
  soft labels from predict_proba, listwise loss, replay buffer, plasticity safeguards).
- local_search_selection_net_v2 with config-flag variants (block removal ladders, unlock/bandit
  rung policies, outer retries, per-phase budget shares, adaptive method order, random ablation).
- selector_monitor: prequential metrics, per-phase oracle-call accounting, per-instance CSV.
- Evaluator instance_ids filter (fixed-instance calibration sets), FED float serialisation fix.
- Configs for synthie/bbbp (10 folds) and proteins10/proteins15, queue file, analysis script.
- SELECTOR_EXPERIMENTS_STATUS.md documents results, variant rationale and next steps;
  lab/selector_experiments/ carries results JSON, monitor CSVs and gzipped logs.
- Thesis v1 checkpoints kept as *.thesis-bak; v2b/v2e/v2f checkpoints for Synthie and PROTEINS.

+++ src/evaluation/future/stages/fed.py
@@ -39,7 +39,7 @@ class FeatureEditDistance(MetricStage):
         # Calculating the average fed considering only the correct instances
         fed_metric = 0.0
         if correct_instances > 0:
-            fed_metric = aggregated_fed/correct_instances
+            fed_metric = float(aggregated_fed/correct_instances)
         
         # Writing the metric value into the explanation and returning the explanation
         self.write_into_explanation(explanation, fed_metric)
